fix: 修复进度通知remaining(usize)与processed(u32)类型不匹配

+++ src-tauri/src/services/upload_scheduler.rs
@@ -97,7 +97,7 @@ impl UploadScheduler {
 
                     let succeeded = progress_qm.tasks_uploaded_in_run();
                     let failed_count = progress_qm.tasks_failed_in_run();
-                    let remaining = pending + uploading;
+                    let remaining = (pending + uploading) as u32;
                     let processed = succeeded + failed_count;
                     let total = processed + remaining;
                     let progress_pct = if total > 0 {
